Show how/where to add scope to your JWT.
#### Attaching Directives To Auto-Generated Queries and Mutations

Since neo4j-graphql.js automatically adds Query and Mutation types to the schema, these auto-generated fields cannot be annotated by the user with directives. To enable authorization on the auto-generated queries and mutations, simply enable the `hasScope` directive and it will be added to the generated CRUD API with the appropriate scope for each operation. For example:
```js
import { makeAugmentedSchema } from "neo4j-graphql-js";

const schema = makeAugmentedSchema({
  tyepDefs,
  config: {
    auth: {
      hasScope: true
    }
  }
})
```
It would be great to have a list of scopes here i.e. read, write, edit, delete and what they would look like  in your JWT. Right now this assumes that the reader will know what those scopes are. I know it may be kind of basic but a full example of this and how it works would be very helpful.